backend/pyproject.toml
[project]
name = "adaptive-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "groq>=1.1.0",
    "motor>=3.7.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
